+++ lib/encoders/qsv.py
@@ -174,7 +174,7 @@ class QsvEncoder(Encoder):
         provides = self.provides()
         return provides.get(encoder, {})
 
-    def stream_args(self, stream_info, stream_id, encoder_name):
+    def stream_args(self, stream_info, stream_id, encoder_name, filter_state=None):
         generic_kwargs = {}
         advanced_kwargs = {}
         encoder_args = []